sistema_dual_alumnos/requirements.txt
streamlit
supabase
pandas
python-dotenv
docxtpl
matplotlib
